+++ src/image.hpp
@@ -5,6 +5,7 @@
 
 Image upsample(const Image& image);
 Image image_from_graphics(const Graphics& graphics, const Palette& palette);
+Image scroll_diagonally(const Image& image);
 
 #endif
